Revert profile move
