+++ cmd/root/bundle.go
@@ -4,23 +4,56 @@ import (
 	"context"
 
 	"github.com/databricks/cli/bundle"
+	"github.com/databricks/cli/bundle/env"
 	"github.com/databricks/cli/bundle/phases"
-	"github.com/databricks/cli/cmd/root/bundleflag"
-	"github.com/databricks/cli/cmd/root/profileflag"
 	"github.com/databricks/cli/libs/diag"
-	"github.com/databricks/cli/libs/env"
+	envlib "github.com/databricks/cli/libs/env"
 	"github.com/spf13/cobra"
+	"golang.org/x/exp/maps"
 )
+
+// getTarget returns the name of the target to operate in.
+func getTarget(cmd *cobra.Command) (value string) {
+	// The command line flag takes precedence.
+	flag := cmd.Flag("target")
+	if flag != nil {
+		value = flag.Value.String()
+		if value != "" {
+			return
+		}
+	}
+
+	oldFlag := cmd.Flag("environment")
+	if oldFlag != nil {
+		value = oldFlag.Value.String()
+		if value != "" {
+			return
+		}
+	}
+
+	// If it's not set, use the environment variable.
+	target, _ := env.Target(cmd.Context())
+	return target
+}
+
+func getProfile(cmd *cobra.Command) (value string) {
+	// The command line flag takes precedence.
+	flag := cmd.Flag("profile")
+	if flag != nil {
+		value = flag.Value.String()
+		if value != "" {
+			return
+		}
+	}
+
+	// If it's not set, use the environment variable.
+	return envlib.Get(cmd.Context(), "DATABRICKS_CONFIG_PROFILE")
+}
 
 // configureProfile applies the profile flag to the bundle.
 func configureProfile(cmd *cobra.Command, b *bundle.Bundle) diag.Diagnostics {
-	profile, ok := profileflag.Value(cmd)
-	if !ok || profile == "" {
-		// If it's not set, use the environment variable.
-		profile, ok = env.Lookup(cmd.Context(), "DATABRICKS_CONFIG_PROFILE")
-	}
-
-	if !ok || profile == "" {
+	profile := getProfile(cmd)
+	if profile == "" {
 		return nil
 	}
 
@@ -33,7 +66,7 @@ func configureProfile(cmd *cobra.Command, b *bundle.Bundle) diag.Diagnostics {
 // configureBundle applies basic mutators to the bundle configures it on the command's context.
 func configureBundle(cmd *cobra.Command, b *bundle.Bundle) (*bundle.Bundle, diag.Diagnostics) {
 	var m bundle.Mutator
-	if target := bundleflag.Target(cmd); target == "" {
+	if target := getTarget(cmd); target == "" {
 		m = phases.LoadDefaultTarget()
 	} else {
 		m = phases.LoadNamedTarget(target)
@@ -79,4 +112,37 @@ func TryConfigureBundle(cmd *cobra.Command) (*bundle.Bundle, diag.Diagnostics) {
 	}
 
 	return configureBundle(cmd, b)
+}
+
+// targetCompletion executes to autocomplete the argument to the target flag.
+func targetCompletion(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+	ctx := cmd.Context()
+	b, err := bundle.MustLoad(ctx)
+	if err != nil {
+		cobra.CompErrorln(err.Error())
+		return nil, cobra.ShellCompDirectiveError
+	}
+
+	// Load bundle but don't select a target (we're completing those).
+	diags := bundle.Apply(ctx, b, phases.Load())
+	if err := diags.Error(); err != nil {
+		cobra.CompErrorln(err.Error())
+		return nil, cobra.ShellCompDirectiveError
+	}
+
+	return maps.Keys(b.Config.Targets), cobra.ShellCompDirectiveDefault
+}
+
+func initTargetFlag(cmd *cobra.Command) {
+	// To operate in the context of a bundle, all commands must take an "target" parameter.
+	cmd.PersistentFlags().StringP("target", "t", "", "bundle target to use (if applicable)")
+	cmd.RegisterFlagCompletionFunc("target", targetCompletion)
+}
+
+// DEPRECATED flag
+func initEnvironmentFlag(cmd *cobra.Command) {
+	// To operate in the context of a bundle, all commands must take an "environment" parameter.
+	cmd.PersistentFlags().StringP("environment", "e", "", "bundle target to use (if applicable)")
+	cmd.PersistentFlags().MarkDeprecated("environment", "use --target flag instead")
+	cmd.RegisterFlagCompletionFunc("environment", targetCompletion)
 }
