--- cmd/root/bundleflag/testdata/invalid/databricks.yml
@@ -1,5 +1,0 @@
-bundle:
-  name: invalid
-
-include:
-  - ./i_dont_exist.yml
