--- cmd/root/bundleflag/testdata/valid/databricks.yml
@@ -1,7 +1,0 @@
-bundle:
-  name: valid
-
-targets:
-  foo: ~
-  bar: ~
-  qux: ~
